refactor: consolidate test helpers to reduce duplication
Move test-helpers.ts from tests/handlers/ to tests/ so both handler
and integration tests share a single makeEvent/makeProvider/makeContext.
Eliminates the duplicated provider factory from text-service.test.ts.
