+++ packages/text-service/tests/handlers/revealed.test.ts
@@ -10,7 +10,7 @@
 
 import { describe, it, expect } from 'vitest';
 import { handleRevealed } from '../../src/handlers/revealed.js';
-import { makeEvent, makeProvider, makeContext } from './test-helpers.js';
+import { makeEvent, makeProvider, makeContext } from '../test-helpers.js';
 
 describe('handleRevealed', () => {
   it('should use direct message field when present', () => {
